% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FonctionsUtiles-package.R
\docType{package}
\name{FonctionsUtiles-package}
\alias{FonctionsUtiles}
\alias{FonctionsUtiles-package}
\title{FonctionsUtiles: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
