% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_block_pls.R
\name{run_block_pls}
\alias{run_block_pls}
\title{Individuals and variables graphic of a multiblock PLS regression}
\usage{
run_block_pls(
  data,
  var_to_select,
  blocks_X,
  block_Y,
  ncomp = 2,
  design = "null",
  group,
  var_color
)
}
\arguments{
\item{data}{The data you want to work on}

\item{var_to_select}{a data.frame with two columns, the first columns contains the names of the variables
used for the regression and must be called "var". The second one contains the names of the blocks assigned to ech variables on the var column.}

\item{blocks_X}{a character vector containing the name of each blocks}

\item{block_Y}{a character vector containing the name of each variables belonging to Y block}

\item{ncomp}{the number of components you want to keep for multiblock PLS regression}

\item{design}{the design matrix of PLS regression (use "null" or "full")}

\item{group}{if you want to shape the points on the individuals graph by a group variable
precise the name of the group variable}

\item{var_color}{a continuous variable to fill the points with a gradient.}
}
\value{
Both individual and variables graphics
}
\description{
Individuals and variables graphic of a multiblock PLS regression
}
\examples{
res <- run_block_pls(
data = df,
var_to_select = var_to_select,
blocks_X = c("block_1","block_2","block_3","block_4"),
block_Y = c("Y_1","Y_2"),
ncomp = 2,
group = "sex",
var_color = "Y_1"
)
}
\author{
Gaspard Martet
}
