% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datasets_to_excel.R
\name{save_datasets_to_excel}
\alias{save_datasets_to_excel}
\title{Save datasets to excel}
\usage{
save_datasets_to_excel(datasets, file_name)
}
\arguments{
\item{datasets}{a list of several datasets}

\item{file_name}{The name of the final dataset you want to save in your working directory}
}
\value{
an excel file saved in your working directory with several windows. One window
by saved dataset
}
\description{
Save datasets to excel
}
\examples{
save_datasets_to_excel(datasets = list("iris" = iris, "mtcars" = mtcars),
file_name = "datasets.xlsx")

}
\author{
Gaspard Martet
}
